+++ core/queue.py
@@ -1,10 +1,9 @@
 import asyncio
-from datetime import datetime
 import traceback
 
 import ujson as json
 
-from core.builtins import Bot, Temp, MessageChain
+from core.builtins import Bot, MessageChain
 from core.logger import Logger
 from core.utils.info import get_all_clients_name
 from core.utils.ip import append_ip, fetch_ip_info
